Implement FTPServer and FTPClient, write tests
Add two interfaces: Server and Client.
Client supports two operations: executeList(), which returns a list of
FileInfo (helper class with name of file/directory and information about
is it directory or not), and executeGet(), which writes all file content
to given OutputStream.

Implemented FTPServer and FTPClient.

Add TestFTPServer, which checks:
1. executeList() with existing directory
2. executeList() with not existing directory
3. executeList() with file, but not directory
4. executeGet() with existing file